Crate verdant_parsers_dynamic

verdant-parsers-dynamic

A parser “collection” for verdant which loads languages at runtime.

See the project overview for more information on all parser collections.

This parser “collection” doesn’t actually include any parsers. Instead, it provides a LanguageLoader that implements [LanguageSet] by loading languages at runtime from a given set of directories, very similar to how tree-sitter-loader does, which is used by the tree-sitter CLI.

Features

docs — Meant to be enabled when building docs

Structs

Lang

A successfully loaded language.

LanguageLoader

A [LanguageSet] implementation that loads languages dynamically at runtime.